[Bronze V] Title: 이상한 기호, Time: 100 ms, Memory: 14336 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -6,9 +6,10 @@ public class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringTokenizer st = new StringTokenizer(br.readLine());
 		
-		int A = Integer.parseInt(st.nextToken());
-		int B = Integer.parseInt(st.nextToken());
-		int result = (A + B) * (A - B);
+		long A = Integer.parseInt(st.nextToken());
+		long B = Integer.parseInt(st.nextToken());
+		
+		long result = (A + B) * (A - B);
 		
 		System.out.println(result);
 	}
